requirements.txt
playwright>=1.40.0
pytest>=7.4.0
pytest-playwright>=0.4.3
pytest-html>=4.1.0
pytest-timeout>=2.2.0
pytest-xdist>=3.5.0
python-dotenv>=1.0.0
allure-pytest>=2.13.0
